Drop exec-based script tests that fight untestable __main__ blocks.
Keep batch JSON parsing, run_deployment subprocess wiring, and the extracted nonempty-layer helper; remove the monkeypatch harness and fake generate_active_space main runs.
